requirements.txt
beautifulsoup4==4.6.0
Flask==0.12.1
PyYAML==3.12
requests==2.14.2
